+++ cli/parser.go
@@ -14,6 +14,52 @@ import (
 
 func GetParser(opts *StandardOptions) *cobra.Command {
 
+	var fingerprintCmd = &cobra.Command{
+		Use:   "fingerprint",
+		Short: "Runs fingerprinting against websites",
+		Long:  "Runs active fingerprinting (wappalyzergo then cvemap)",
+		Args:  cobra.MinimumNArgs(0),
+		Run: func(cmd *cobra.Command, args []string) {
+
+			if opts.OutDirName == "" && opts.TargetFilePath == "" {
+				fmt.Println("[!] Error: target cannot be empty. Please run with --target [something] or --file [something]")
+				fmt.Println("\nIf you're confused, feel free to use --help option. :)~")
+				os.Exit(1)
+			}
+
+			helper.CheckProxy(opts.Proxy)
+			helper.DisplayNetInfo()
+
+			if opts.OutDirName == "" || !helper.Exists(opts.OutDirName) {
+				println("[!] collected assets will be sent to current working directory")
+			}
+
+			scanOpts := scan.ScanOpts{}
+			scanOpts.SetDomain(opts.OutDirName)
+			scanOpts.SetScanPath(opts.OutDirName)
+			scanOpts.SetProxy(opts.Proxy)
+			scanOpts.SetDryRun(opts.RunDry)
+			scanOpts.SetRateLimit(opts.RateLimit)
+			scanOpts.SetWordlistPath(opts.WordlistPath)
+			scanOpts.SetForceInsecure(opts.UseHttpInsecure)
+			scanOpts.SetNoGobuster(opts.NoGobuster)
+
+			if opts.TargetFilePath == "" {
+				scanOpts.Run()
+				return
+			}
+
+			scanner := helper.LoadTargetFile(opts.TargetFilePath)
+			defer scanner.Close()
+
+			for scanner.Scan() {
+				targetDomain := scanner.Text()
+				scanOpts.SetDomain(targetDomain)
+				scanOpts.Fingerprint()
+			}
+		},
+	}
+
 	var scanCmd = &cobra.Command{
 		Use:   "scan",
 		Short: "Run active scanning tools to perform enumeration",
@@ -149,11 +195,17 @@ func GetParser(opts *StandardOptions) *cobra.Command {
 	scanCmd.Flags().StringVarP(&opts.OutDirName, "dir-name", "d", defaults.OutDirName, "Outfile directory name (if no target-file is specified, will also be target domain)")
 	scanCmd.Flags().StringVarP(&opts.Proxy, "proxy", "p", defaults.Proxy, "Proxy URL (used for the tools supporting it. Other will prompt a warning msg)")
 	scanCmd.Flags().BoolVarP(&opts.RunDry, "dry", "", defaults.RunDry, "Run a dry-run (test mode)")
-	scanCmd.Flags().BoolVarP(&opts.NoGobuster, "no-dirbusting", "", defaults.NoGobuster, "Disable directory bruteforce (sometimes you don't need it yk...)")
+	scanCmd.Flags().BoolVarP(&opts.NoGobuster, "disable-dirbusting", "", defaults.NoGobuster, "Disable directory bruteforce (sometimes you don't need it yk...)")
 	scanCmd.Flags().BoolVarP(&opts.UseHttpInsecure, "insecure", "k", defaults.UseHttpInsecure, "Ignore SSL warnings and force http")
 	scanCmd.Flags().Int32VarP(&opts.RateLimit, "rate-limit", "r", defaults.RateLimit, "Requests rate-limit (used for the tools supporting it. Other will prompt a warning msg)")
 	scanCmd.Flags().StringVarP(&opts.WordlistPath, "wordlist", "w", defaults.WordlistPath, "Wordlist to use")
 	scanCmd.Flags().StringVarP(&opts.TargetFilePath, "file", "f", defaults.TargetFilePath, "File containing list of targets (should be a list of IP Addresses or domains)")
+
+	rootCmd.AddCommand(fingerprintCmd)
+	fingerprintCmd.Flags().StringVarP(&opts.OutDirName, "dir-name", "d", defaults.OutDirName, "Outfile directory name (if no target-file is specified, will also be target domain)")
+	fingerprintCmd.Flags().StringVarP(&opts.Proxy, "proxy", "p", defaults.Proxy, "Proxy URL (used for the tools supporting it. Other will prompt a warning msg)")
+	fingerprintCmd.Flags().BoolVarP(&opts.UseHttpInsecure, "insecure", "k", defaults.UseHttpInsecure, "Ignore SSL warnings and force http")
+	fingerprintCmd.Flags().StringVarP(&opts.TargetFilePath, "file", "f", defaults.TargetFilePath, "File containing list of targets (should be a list of IP Addresses or domains)")
 
 	return rootCmd
 }