+++ cli/types.go
@@ -3,8 +3,9 @@ package cli
 type StandardOptions struct {
 
 	// Pentest target, will be the name of the directories created, for instance
-	// if no target file is specified for OSINT command, will use that instead
-	OutDirName string
+	// if no target file is specified for OSINT command, will use that instead.
+	// Can be Outfile or OutDir
+	OutName string
 
 	// Run a dry-run (test mode)
 	RunDry bool
@@ -30,7 +31,7 @@ type StandardOptions struct {
 
 func GetDefaultOptions() StandardOptions {
 	opts := StandardOptions{
-		OutDirName:      "",
+		OutName:         "",
 		RunDry:          false,
 		Proxy:           "",
 		UseHttpInsecure: false,