+++ Changelog.md
@@ -1,3 +1,6 @@
+## v0.0.5
+
+
 ## v0.0.4
 
 - feat(scan): saving wappalyzer results to file ~ @djnnvx